- It has been moved the ApiController to the root controllers directory - It has been moved the auth routes outside the api/v1 scope.
