+++ app/controllers/api_controller.rb
@@ -1,0 +1,10 @@
+class ApiController < ApplicationController
+  respond_to :json
+  rescue_from ActiveRecord::RecordNotFound, with: :not_found
+
+  private
+
+  def not_found
+    render json: { error: "Record not found with id #{params[:id]}" }, status: :not_found
+  end
+end
